requirements.txt
flask==2.0.2
requests==2.25.0
redis==4.2.2
PyYAML==5.3.1